Replicated changes from pulseMiddle() to pulseRight() and pulseLeft()
Lets user know when mouse can turn to make it easier to exxecute a turn. NOTE: if a single sensor gets very blocked as it nears an opening, measurements become really off and reach 3000cm. Since the actions are triggered in an if statements if only needs to happen once to mess everything up. Some ways to circumvent this is to have the if statement execute given a range rather than a single value comparison. Furthermore depending on the stepper motor speed the pulse rate can be played with to prevent this from happening.